Implemented the GRU_teacher model and fixed tests
